Private API

Xfoil.XfoilGlobals — Type

XfoilGlobals

Holds XFOIL global variables

Xfoil.get_globals — Function

get_globals()

Wraps the global variables in XFOIL in a struct

Xfoil.do_percussive_maintenance — Function

do_percussive_maintenance(x, y, alpha, re, iter, npan, ncrit)

Attempts to converge previously unconverged XFOIL solutions by perturbing the starting point for the XFOIL solve. Returns cl, cd, cdp, cm, converged

Xfoil.get_globals_cs — Function

get_globals_cs()

get_globals for the complex step enabled version of XFOIL.

Xfoil.do_percussive_maintenance_cs — Function

do_percussive_maintenance_cs(x, y, alpha, re, iter, npan, ncrit)

do_percussive_maintenance for the complex step enabled version of XFOIL.